remove controller from soltaro, move tests from soltaro to new controller
