--- io.openems.edge.battery.soltaro/src/io/openems/edge/battery/soltaro/controller/IState.java
@@ -1,30 +1,0 @@
-package io.openems.edge.battery.soltaro.controller;
-
-import io.openems.common.exceptions.OpenemsError.OpenemsNamedException;
-
-public interface IState {
-
-	/**
-	 * Returns the corresponding state.
-	 * 
-	 * @return the state
-	 */
-	State getState();
-
-	/**
-	 * Depending on the circumstances the next state according to the state machine
-	 * is returned.
-	 * 
-	 * @return the next state
-	 */
-	State getNextState();
-
-	/**
-	 * in this method everything should be executed what there is to do in this
-	 * state.
-	 * 
-	 * @throws OpenemsNamedException on error
-	 */
-	void act() throws OpenemsNamedException;
-
-}
