+++ io.openems.edge.controller.battery.batteryprotection/test/io/openems/edge/controller/battery/batteryprotection/helper/DummyBattery.java
@@ -1,12 +1,10 @@
-package io.openems.edge.battery.soltaro.controller.helper;
+package io.openems.edge.controller.battery.batteryprotection.helper;
 
 import io.openems.edge.battery.api.Battery;
-import io.openems.edge.battery.soltaro.ChargeIndication;
-import io.openems.edge.battery.soltaro.SoltaroBattery;
 import io.openems.edge.common.component.AbstractOpenemsComponent;
 import io.openems.edge.common.component.OpenemsComponent;
 
-public class DummyBattery extends AbstractOpenemsComponent implements SoltaroBattery {
+public class DummyBattery extends AbstractOpenemsComponent implements Battery {
 
 	public static int DEFAULT_SOC = 50;
 	public static int DEFAULT_MIN_CELL_VOLTAGE = 3280;
@@ -14,14 +12,14 @@ public class DummyBattery extends AbstractOpenemsComponent implements SoltaroBat
 	public static int DEFAULT_MIN_CELL_TEMPERATURE = 25;
 	public static int DEFAULT_MAX_CELL_TEMPERATURE = 33;
 
-	private static final ChargeIndication DEFAULT_CHARGE_INDICATION = ChargeIndication.STANDBY;
+//	private static final ChargeIndication DEFAULT_CHARGE_INDICATION = ChargeIndication.STANDBY;
 
 	protected DummyBattery(//
 	) { //
 		super(//
 				OpenemsComponent.ChannelId.values(), //
-				Battery.ChannelId.values(), //
-				SoltaroBattery.ChannelId.values() //
+				Battery.ChannelId.values()
+//				, SoltaroBattery.ChannelId.values() //
 		);
 
 //		getChargeIndication().onSetNextValue( v -> { 
@@ -42,7 +40,7 @@ public class DummyBattery extends AbstractOpenemsComponent implements SoltaroBat
 		setMinimalCellTemperature(DEFAULT_MIN_CELL_TEMPERATURE);
 		setMaximalCellTemperature(DEFAULT_MAX_CELL_TEMPERATURE);
 		setSoc(DEFAULT_SOC);
-		setChargeIndication(DEFAULT_CHARGE_INDICATION);
+//		setChargeIndication(DEFAULT_CHARGE_INDICATION);
 	}
 
 	public void setMinimalCellVoltage(int minimalCellVoltage) {
@@ -85,15 +83,15 @@ public class DummyBattery extends AbstractOpenemsComponent implements SoltaroBat
 		this.getMaxCellTemperature().nextProcessImage();
 	}
 
-	public void setChargeIndication(ChargeIndication chargeIndication) {
-		this.getChargeIndication().setNextValue(chargeIndication);
-		this.getChargeIndication().nextProcessImage();
-	}
-
-	public void setChargeIndicationToUndefined() {
-		this.getChargeIndication().setNextValue(null);
-		this.getChargeIndication().nextProcessImage();
-	}
+//	public void setChargeIndication(ChargeIndication chargeIndication) {
+//		this.getChargeIndication().setNextValue(chargeIndication);
+//		this.getChargeIndication().nextProcessImage();
+//	}
+//
+//	public void setChargeIndicationToUndefined() {
+//		this.getChargeIndication().setNextValue(null);
+//		this.getChargeIndication().nextProcessImage();
+//	}
 
 	public void setSoc(int soc) {
 		this.getSoc().setNextValue(soc);
